fix(dsl_compiler): add IrType::{I8, U8, I16, U16} variants
Milestone 2 flagged SlowApplied.factor_q8 as a workaround (wrote i32
because u16 resolved to IrType::Named). Now u16 is a real primitive.
Covers every stdlib integer width the engine's niche-ID types + hot
SoA fields actually use.

Changes:
- IrType gains I8/U8/I16/U16 variants alongside I32/U32/I64/U64
- Resolver seeds "i8", "u8", "i16", "u16" in stdlib_types
- schema_hash canonical bytes assigned 0x0a-0x0d for the new widths
- emit_rust renders them via the standard rustc spelling; Copy set
  extended to include them
- emit_python maps all four to Python `int` (same bucket as I32/U64)
- assets/sim/events.sim: factor_q8 now u16 (matches engine SoA)
- Regenerated engine_rules output + reblessed IR goldens
- EVENT_HASH: ea2436c6...b794291 (was cb634dda...82453b)

All 27 dsl_compiler tests pass; workspace builds clean; --check
passes against regenerated output.

+++ crates/engine_rules/src/events/slow_applied.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 pub struct SlowApplied {
     #[serde(with = "crate::id_serde::agent_id")]
     pub target: AgentId,
-    pub factor_q8: i32,
+    pub factor_q8: u16,
     pub remaining_ticks: u32,
     pub tick: u64,
 }
